requirements.txt
fastapi==0.109.1
python-multipart==0.0.6
uvicorn==0.24.0
Pillow==10.1.0
numpy<2.0.0
opencv-python==4.8.1.78
easyocr==1.7.1
torch>=2.0.0
torchvision>=0.15.0
